api testing, further dev

+++ scripts2/transpile_spirv_glsl.sh
@@ -1,30 +1,34 @@
 #!/usr/bin/env bash
 
 ### create directory
-mkdir -p "../src/main/resources/shaders/hlsl/"
+mkdir -p "../src/main/resources/shaders/glsl/"
 
 ### cleanup
-for file in ../src/main/resources/shaders/hlsl/*.vert.hlsl; do
-  echo "removing ${file}"
-  rm ${file}
+for file in ../src/main/resources/shaders/glsl/*.vert.glsl; do
+  if [ -f ${file} ]; then
+    echo "removing ${file}"
+    rm ${file}
+  fi
 done
 
-for file in ../src/main/resources/shaders/hlsl/*.frag.hlsl; do
-  echo "removing ${file}"
-  rm ${file}
+for file in ../src/main/resources/shaders/glsl/*.frag.glsl; do
+  if [ -f ${file} ]; then
+    echo "removing ${file}"
+    rm ${file}
+  fi
 done
 
 ### shader gen
 echo "generating shaders..."
 
 for file in ../src/main/resources/shaders/spirv/*.vert.spv; do
-  echo "transpiling ${file} to hlsl"
+  echo "transpiling ${file} to glsl"
   baseName=$(basename ${file} ".vert.spv")
-  spirv-cross ${file} --hlsl --output ../src/main/resources/shaders/hlsl/${baseName}.vert.hlsl
+  spirv-cross ${file} --version 450 --no-es --output ../src/main/resources/shaders/glsl/${baseName}.vert.glsl
 done
 
 for file in ../src/main/resources/shaders/spirv/*.frag.spv; do
-  echo "transpiling ${file} to hlsl"
+  echo "transpiling ${file} to glsl"
   baseName=$(basename ${file} ".frag.spv")
-  spirv-cross ${file} --hlsl --output ../src/main/resources/shaders/hlsl/${baseName}.frag.hlsl
+  spirv-cross ${file} --version 450 --no-es --output ../src/main/resources/shaders/glsl/${baseName}.frag.glsl
 done
