+++ scripts2/transpile_spirv_glsl_es.sh
@@ -1,30 +1,34 @@
 #!/usr/bin/env bash
 
 ### create directory
-mkdir -p "../src/main/resources/shaders/glsl/"
+mkdir -p "../src/main/resources/shaders/glsl-es/"
 
 ### cleanup
-for file in ../src/main/resources/shaders/glsl/*.vert.glsl; do
-  echo "removing ${file}"
-  rm ${file}
+for file in ../src/main/resources/shaders/glsl-es/*.vert.es.glsl; do
+  if [ -f ${file} ]; then
+    echo "removing ${file}"
+    rm ${file}
+  fi
 done
 
-for file in ../src/main/resources/shaders/glsl/*.frag.glsl; do
-  echo "removing ${file}"
-  rm ${file}
+for file in ../src/main/resources/shaders/glsl-es/*.frag.es.glsl; do
+  if [ -f ${file} ]; then
+    echo "removing ${file}"
+    rm ${file}
+  fi
 done
 
 ### shader gen
 echo "generating shaders..."
 
 for file in ../src/main/resources/shaders/spirv/*.vert.spv; do
-  echo "transpiling ${file} to glsl"
+  echo "transpiling ${file} to glsl-es"
   baseName=$(basename ${file} ".vert.spv")
-  spirv-cross ${file} --version 450 --no-es --output ../src/main/resources/shaders/glsl/${baseName}.vert.glsl
+  spirv-cross ${file} --version 310 --es --output ../src/main/resources/shaders/glsl-es/${baseName}.vert.es.glsl
 done
 
 for file in ../src/main/resources/shaders/spirv/*.frag.spv; do
-  echo "transpiling ${file} to glsl"
+  echo "transpiling ${file} to glsl-es"
   baseName=$(basename ${file} ".frag.spv")
-  spirv-cross ${file} --version 450 --no-es --output ../src/main/resources/shaders/glsl/${baseName}.vert.glsl
+  spirv-cross ${file} --version 310 --es --output ../src/main/resources/shaders/glsl-es/${baseName}.frag.es.glsl
 done
